# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Music Service",
...
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "YouTube Music API Service",
      "url": "https://github.com/your-repo"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/search/": {
      "get": {
        "tags": [
          "search",
          "search"
        ],
        "summary": "Search music content",
        "description": "Busca contenido musical en YouTube Music: canciones, videos, álbumes, artistas y playlists.",
        "operationId": "search_music_api_v1_search__get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Query de búsqueda",
              "examples": [
                "cumbia peruana"
              ],
              "title": "Q"
            },
            "description": "Query de búsqueda"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro: songs, videos, albums, artists, playlists",
              "examples": [
                "songs"
              ],
              "title": "Filter"
            },
            "description": "Filtro: songs, videos, albums, artists, playlists"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scope de búsqueda",
              "title": "Scope"
            },
            "description": "Scope de búsqueda"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Número de resultados",
              "default": 10,
              "title": "Limit"
            },
            "description": "Número de resultados"
          },
          {
            "name": "start_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Índice inicial para paginación",
              "default": 0,
              "title": "Start Index"
            },
            "description": "Índice inicial para paginación"
          },
          {
            "name": "ignore_spelling",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Ignorar sugerencias de ortografía",
              "default": false,
              "title": "Ignore Spelling"
            },
            "description": "Ignorar sugerencias de ortografía"
          },
          {
            "name": "include_stream_urls",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Incluir stream URLs y mejores thumbnails para songs/videos",
              "default": true,
              "title": "Include Stream Urls"
            },
            "description": "Incluir stream URLs y mejores thumbnails para songs/videos"
          }
        ],
        "responses": {
          "200": {
            "description": "Búsqueda exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Music Api V1 Search  Get"
                },
                "example": {
                  "results": [
                    {
                      "videoId": "rMbATaj7Il8",
                      "title": "Song Title",
                      "artists": [
                        {
                          "name": "Artist"
                        }
                      ],
                      "album": {
                        "name": "Album"
                      },
                      "stream_url": "https://...",
                      "thumbnail": "https://..."
                    }
                  ],
                  "query": "search query"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/suggestions": {
      "get": {
        "tags": [
          "search",
          "search"
        ],
        "summary": "Get search suggestions",
        "description": "Obtiene sugerencias de búsqueda basadas en el query parcial.",
        "operationId": "get_search_suggestions_api_v1_search_suggestions_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Query parcial para obtener sugerencias",
              "examples": [
                "cumb"
              ],
              "title": "Q"
            },
            "description": "Query parcial para obtener sugerencias"
          }
        ],
        "responses": {
          "200": {
            "description": "Sugerencias obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Search Suggestions Api V1 Search Suggestions Get"
                },
                "example": {
                  "suggestions": [
                    "cumbia peruana",
                    "cumbia colombiana",
                    "cumbia villera"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "search",
          "search"
        ],
        "summary": "Remove search suggestions",
        "description": "Elimina una sugerencia de búsqueda del historial.",
        "operationId": "remove_search_suggestions_api_v1_search_suggestions_delete",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Query a eliminar de las sugerencias",
              "examples": [
                "cumbia"
              ],
              "title": "Q"
            },
            "description": "Query a eliminar de las sugerencias"
          }
        ],
        "responses": {
          "200": {
            "description": "Sugerencia eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Search Suggestions Api V1 Search Suggestions Delete"
                },
                "example": {
                  "success": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "boolean",
            "title": "Error",
            "description": "Always True for error responses",
            "default": true
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Machine-readable error code",
            "examples": [
              "VALIDATION_ERROR",
              "AUTHENTICATION_ERROR",
              "RATE_LIMIT_ERROR",
              "NOT_FOUND",
              "EXTERNAL_SERVICE_ERROR",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message in Spanish",
            "examples": [
              "Error de validación en los parámetros de la petición",
              "Error de autenticación con YouTube Music",
              "Límite de peticiones excedido. Intenta más tarde.",
              "Recurso no encontrado",
              "Error en el servicio de YouTube Music",
              "Servicio temporalmente no disponible"
            ]
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional error details (structure varies by error type)"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema for all API errors.\n\nThis schema is used consistently across all endpoints to provide\na predictable error format for clients.\n\nAttributes:\n    error: Always True for error responses.\n    error_code: Machine-readable error code for programmatic handling.\n    message: Human-readable error message in Spanish.\n    details: Optional additional details about the error."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}